% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{animals}
\alias{animals}
\title{Small example data set for self-organized artificial neural networks}
\format{A data frame with 10 rows and 6 variables describing 10 different
  animals with 6 feature vectors that are either present (1) or absent (0).}
\source{
Knight, K. (1990). Connectionist ideas and algorithms.
  \emph{Communications of the ACM}, \emph{33}(11), 59–74.
}
\usage{
animals
}
\description{
A small dataset describing 10 animals represented by 6 features that are
either present (1) or absent (0) for demonstrating self-organized learning in
artificial neural networks.
}
\keyword{datasets}
