% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leabRa.R
\docType{package}
\name{leabRa}
\alias{leabRa}
\alias{leabRa-package}
\title{leabRa: A package for biologically realistic neural networks based on leabra}
\description{
The leabRa package provides three classes to construct artifical neural
networks: \link{unit}, \link{layer} and \link{network}.
}
\details{
Note that the classes in this package are R6 classes.

For further information check out the vignette.
}
