% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leabRa.R
\docType{package}
\name{leabRa}
\alias{leabRa}
\alias{leabRa-package}
\title{leabRa: A package for biologically realistic neural networks based on Leabra}
\description{
The Leabra package provides three classes to construct artificial neural
networks: \link{unit}, \link{layer} and \link{network}.
}
\details{
Note that the classes in this package are \code{\link{R6Class}} classes.

For further information check out the vignette with vignette("leabRa").
}
